+++ apps/html-storybook/vite.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'vite'
 // https://vitejs.dev/config/
 export default defineConfig({
   publicDir: './../public/',
+  // because pnpm and stories are in node_modules
   resolve: {
     preserveSymlinks: true
   },
