+++ apps/vue-storybook/.storybook/main.ts
@@ -2,11 +2,8 @@
 import remarkGfm from 'remark-gfm'
 const config = {
   stories: [
-    './../stories/**/*.stories.@(js|jsx|ts|tsx)',
-    // './../stories/**/*.docs.mdx',
-    // './../docs/**/*.stories.@(js|jsx|ts|tsx)',
-    // './../docs/**/*.docs.mdx',
-    './../../../packages/vue/src/**/*.stories.@(js|jsx|ts|tsx)'
+    './../node_modules/@explorer-1/vue/src/**/*.stories.@(js|jsx|ts|tsx)',
+    './../node_modules/@explorer-1/vue/src/**/*.docs.mdx'
   ],
   addons: [
     {
