--- apps/vue-storybook/stories/components/BlockDialog.stories.mdx
@@ -1,82 +1,0 @@
-import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import BlockDialog from '@explorer-1/vue/src/components/BlockDialog/BlockDialog.vue'
-
-<Meta
-  title="Components/Blocks/BlockDialog"
-  component={BlockDialog}
-  parameters={{ viewMode: 'docs' }}
-  argTypes={{
-    overlayClose: { control: { type: 'boolean' } }
-  }}
-/>
-
-export const BlockDialogTemplate = (args) => ({
-  props: Object.keys(args),
-  components: { BlockDialog },
-  template: `
-  <div>
-    <button type="button" data-a11y-dialog-show="dialog-id">Open the dialog</button>
-    <BlockDialog :dialog-id="dialogId" :heading-id="headingId" dialog-box-class="max-w-2xl">
-      <div class="p-3 sm:p-5">
-        <h2 id="heading-id">Dialog title</h2>
-        <hr
-          aria-hidden="true"
-          class="bg-gray-light-mid w-full h-2px border-0 my-3"
-        />
-        <div class="dialog-main py-3 max-h-64 overflow-auto relative">
-          <p class="py-2">
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-            eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-            enim ad minim veniam, quis nostrud exercitation ullamco laboris
-            nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor
-            in reprehenderit in voluptate velit esse cillum dolore eu fugiat
-            nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-            sunt in culpa qui officia deserunt mollit anim id est laborum.
-          </p>
-        </div>
-        <div class="my-3">
-          <a
-            href="#"
-            target=""
-            rel=""
-            class="group cursor-pointer -default underline text-theme-color can-hover:hover:text-theme-color-hover"
-          >
-            <span class="break-words">Explore</span>
-          </a>
-        </div>
-      </div>
-    </BlockDialog>
-  </div>
-`
-})
-
-# BlockDialog
-
-This component is an accessible dialog window that utilizes the efficient and lightweight `a11y-dialog` package.
-It adheres to the [Dialog (Modal) pattern](https://www.w3.org/WAI/ARIA/apg/patterns/dialog-modal/) outlined in the Accessible Rich Internet Applications (ARIA) Authoring Practices Guide.
-
-Additionally, it is unopinionated with styling.
-
-<Canvas className="sbdocs-preview-full-width">
-  <Story
-    name="Default"
-    args={{
-      dialogId: 'dialog-id',
-      headingId: 'heading-id'
-    }}
-  >
-    {BlockDialogTemplate.bind({})}
-  </Story>
-</Canvas>
-
-### Basic Usage
-
-To trigger dialog opening, use the `data-a11y-dialog-show` attribute with the value set to the `id` of the dialog element you want to open.
-
-```
-<button type="button" data-a11y-dialog-show="your-dialog-id">
-  Open the dialog
-</button>
-```
-
-See [all documentation in the repository](https://github.com/KittyGiraudel/a11y-dialog).
