+++ apps/vue-storybook/tailwind.config.ts
@@ -11,9 +11,8 @@ import explorer1Config from '@explorer-1/common/tailwind.config'
 export default {
   ...explorer1Config,
   content: [
-    'stories/**/*',
-    'docs/**/*',
-    './node_modules/@explorer-1/vue/src/components/**/*.vue',
-    './node_modules/@explorer-1/vue/src/templates/**/*.vue'
+    './node_modules/@explorer-1/vue/src/docs/**/*',
+    './node_modules/@explorer-1/vue/src/components/**/*',
+    './node_modules/@explorer-1/vue/src/templates/**/*'
   ]
 }
