+++ packages/vue/src/components/BaseCheckboxGroup/BaseCheckboxGroup.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import BaseCheckboxGroup from '@explorer-1/vue/src/components/BaseCheckboxGroup/BaseCheckboxGroup.vue'
+import BaseCheckboxGroup from './BaseCheckboxGroup.vue'
 
 <Meta
   title="Components/Base/BaseCheckboxGroup"
