+++ packages/vue/src/components/BaseRadioGroup/BaseRadioGroup.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import BaseRadioGroup from '@explorer-1/vue/src/components/BaseRadioGroup/BaseRadioGroup.vue'
+import BaseRadioGroup from './BaseRadioGroup.vue'
 
 <Meta
   title="Components/Base/BaseRadioGroup"
