+++ packages/vue/src/components/BaseSwimlane/BaseSwimlane.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import BaseSwimlane from '@explorer-1/vue/src/components/BaseSwimlane/BaseSwimlane.vue'
+import BaseSwimlane from './BaseSwimlane.vue'
 
 <Meta
   title="Components/Base/BaseSwimlane"
