+++ packages/vue/src/components/Icons/Icons.stories.ts
@@ -1,56 +1,56 @@
-import IconArrow from '@explorer-1/vue/src/components/Icons/IconArrow.vue'
-import Icon360 from '@explorer-1/vue/src/components/Icons/Icon360.vue'
-import IconArrows from '@explorer-1/vue/src/components/Icons/IconArrows.vue'
-import IconAudio from '@explorer-1/vue/src/components/Icons/IconAudio.vue'
-import IconBook from '@explorer-1/vue/src/components/Icons/IconBook.vue'
-import IconBookUser from '@explorer-1/vue/src/components/Icons/IconBookUser.vue'
-import IconBriefcase from '@explorer-1/vue/src/components/Icons/IconBriefcase.vue'
-import IconCalendar from '@explorer-1/vue/src/components/Icons/IconCalendar.vue'
-import IconCaret from '@explorer-1/vue/src/components/Icons/IconCaret.vue'
-import IconCheckbox from '@explorer-1/vue/src/components/Icons/IconCheckbox.vue'
-import IconCheckboxSelected from '@explorer-1/vue/src/components/Icons/IconCheckboxSelected.vue'
-import IconClose from '@explorer-1/vue/src/components/Icons/IconClose.vue'
-import IconCloseLightbox from '@explorer-1/vue/src/components/Icons/IconCloseLightbox.vue'
-import IconDownload from '@explorer-1/vue/src/components/Icons/IconDownload.vue'
-import IconDropdown from '@explorer-1/vue/src/components/Icons/IconDropdown.vue'
-import IconEnvelope from '@explorer-1/vue/src/components/Icons/IconEnvelope.vue'
-import IconExpand from '@explorer-1/vue/src/components/Icons/IconExpand.vue'
-import IconExternal from '@explorer-1/vue/src/components/Icons/IconExternal.vue'
-import IconFilter from '@explorer-1/vue/src/components/Icons/IconFilter.vue'
-import IconForwardTen from '@explorer-1/vue/src/components/Icons/IconForwardTen.vue'
-import IconGrid from '@explorer-1/vue/src/components/Icons/IconGrid.vue'
-import IconHand from '@explorer-1/vue/src/components/Icons/IconHand.vue'
-import IconInfo from '@explorer-1/vue/src/components/Icons/IconInfo.vue'
-import IconLink from '@explorer-1/vue/src/components/Icons/IconLink.vue'
-import IconList from '@explorer-1/vue/src/components/Icons/IconList.vue'
-import IconLocation from '@explorer-1/vue/src/components/Icons/IconLocation.vue'
-import IconMagnifyingGlass from '@explorer-1/vue/src/components/Icons/IconMagnifyingGlass.vue'
-import IconMedal from '@explorer-1/vue/src/components/Icons/IconMedal.vue'
-import IconMenu from '@explorer-1/vue/src/components/Icons/IconMenu.vue'
-import IconMinus from '@explorer-1/vue/src/components/Icons/IconMinus.vue'
-import IconMute from '@explorer-1/vue/src/components/Icons/IconMute.vue'
-import IconNext from '@explorer-1/vue/src/components/Icons/IconNext.vue'
-import IconPause from '@explorer-1/vue/src/components/Icons/IconPause.vue'
-import IconPlay from '@explorer-1/vue/src/components/Icons/IconPlay.vue'
-import IconPlus from '@explorer-1/vue/src/components/Icons/IconPlus.vue'
-import IconPrev from '@explorer-1/vue/src/components/Icons/IconPrev.vue'
-import IconRewindTen from '@explorer-1/vue/src/components/Icons/IconRewindTen.vue'
-import IconSearch from '@explorer-1/vue/src/components/Icons/IconSearch.vue'
-import IconShare from '@explorer-1/vue/src/components/Icons/IconShare.vue'
-import IconSlideshow from '@explorer-1/vue/src/components/Icons/IconSlideshow.vue'
-import IconSocialEmail from '@explorer-1/vue/src/components/Icons/IconSocialEmail.vue'
-import IconSocialFacebook from '@explorer-1/vue/src/components/Icons/IconSocialFacebook.vue'
-import IconSocialGoogleClassroom from '@explorer-1/vue/src/components/Icons/IconSocialGoogleClassroom.vue'
-import IconSocialInstagram from '@explorer-1/vue/src/components/Icons/IconSocialInstagram.vue'
-import IconSocialPinterest from '@explorer-1/vue/src/components/Icons/IconSocialPinterest.vue'
-import IconSocialReddit from '@explorer-1/vue/src/components/Icons/IconSocialReddit.vue'
-import IconSocialTwitter from '@explorer-1/vue/src/components/Icons/IconSocialTwitter.vue'
-import IconSocialYoutube from '@explorer-1/vue/src/components/Icons/IconSocialYoutube.vue'
-import IconStop from '@explorer-1/vue/src/components/Icons/IconStop.vue'
-import IconTime from '@explorer-1/vue/src/components/Icons/IconTime.vue'
-import IconUniversity from '@explorer-1/vue/src/components/Icons/IconUniversity.vue'
-import IconUser from '@explorer-1/vue/src/components/Icons/IconUser.vue'
-import IconVolume from '@explorer-1/vue/src/components/Icons/IconVolume.vue'
+import IconArrow from './IconArrow.vue'
+import Icon360 from './Icon360.vue'
+import IconArrows from './IconArrows.vue'
+import IconAudio from './IconAudio.vue'
+import IconBook from './IconBook.vue'
+import IconBookUser from './IconBookUser.vue'
+import IconBriefcase from './IconBriefcase.vue'
+import IconCalendar from './IconCalendar.vue'
+import IconCaret from './IconCaret.vue'
+import IconCheckbox from './IconCheckbox.vue'
+import IconCheckboxSelected from './IconCheckboxSelected.vue'
+import IconClose from './IconClose.vue'
+import IconCloseLightbox from './IconCloseLightbox.vue'
+import IconDownload from './IconDownload.vue'
+import IconDropdown from './IconDropdown.vue'
+import IconEnvelope from './IconEnvelope.vue'
+import IconExpand from './IconExpand.vue'
+import IconExternal from './IconExternal.vue'
+import IconFilter from './IconFilter.vue'
+import IconForwardTen from './IconForwardTen.vue'
+import IconGrid from './IconGrid.vue'
+import IconHand from './IconHand.vue'
+import IconInfo from './IconInfo.vue'
+import IconLink from './IconLink.vue'
+import IconList from './IconList.vue'
+import IconLocation from './IconLocation.vue'
+import IconMagnifyingGlass from './IconMagnifyingGlass.vue'
+import IconMedal from './IconMedal.vue'
+import IconMenu from './IconMenu.vue'
+import IconMinus from './IconMinus.vue'
+import IconMute from './IconMute.vue'
+import IconNext from './IconNext.vue'
+import IconPause from './IconPause.vue'
+import IconPlay from './IconPlay.vue'
+import IconPlus from './IconPlus.vue'
+import IconPrev from './IconPrev.vue'
+import IconRewindTen from './IconRewindTen.vue'
+import IconSearch from './IconSearch.vue'
+import IconShare from './IconShare.vue'
+import IconSlideshow from './IconSlideshow.vue'
+import IconSocialEmail from './IconSocialEmail.vue'
+import IconSocialFacebook from './IconSocialFacebook.vue'
+import IconSocialGoogleClassroom from './IconSocialGoogleClassroom.vue'
+import IconSocialInstagram from './IconSocialInstagram.vue'
+import IconSocialPinterest from './IconSocialPinterest.vue'
+import IconSocialReddit from './IconSocialReddit.vue'
+import IconSocialTwitter from './IconSocialTwitter.vue'
+import IconSocialYoutube from './IconSocialYoutube.vue'
+import IconStop from './IconStop.vue'
+import IconTime from './IconTime.vue'
+import IconUniversity from './IconUniversity.vue'
+import IconUser from './IconUser.vue'
+import IconVolume from './IconVolume.vue'
 
 export default {
   title: 'Foundation/Icons',
