+++ packages/vue/src/components/MixinCarousel/MixinCarousel.docs.mdx
