+++ packages/vue/src/docs/foundation/FontVariants.vue
@@ -1,0 +1,80 @@
+<template>
+  <div class="Fonts md:grid-cols-2 grid gap-2 mt-4">
+    <div>
+      <h3
+        id="primary"
+        class="inline-block align-top mr-4 font-primary"
+      >
+        Metropolis
+        <br />
+        <code class="text-base">.font-primary</code>
+      </h3>
+      <ol class="list-none inline-block">
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-light"> Metropolis light </span>
+            <code>.font-light</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-normal"> Metropolis normal </span>
+            <code>.font-normal</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-medium"> Metropolis medium </span>
+            <code>.font-medium</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-semibold"> Metropolis semi bold </span>
+            <code>.font-semibold</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-bold">Metropolis bold</span>
+            <code>.font-bold</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-primary font-extrabold"> Metropolis extra bold </span>
+            <code>.font-extrabold</code>
+          </span>
+        </li>
+      </ol>
+    </div>
+    <div>
+      <h3
+        id="secondary"
+        class="inline-block align-top mr-4 font-secondary"
+      >
+        Archivo Narrow
+        <br />
+        <code class="text-base">.font-secondary</code>
+      </h3>
+      <ol class="list-none inline-block">
+        <li></li>
+        <li></li>
+        <li></li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-secondary font-semibold"> Archivo Narrow semi bold </span>
+            <code>.font-semibold</code>
+          </span>
+        </li>
+        <li>
+          <span class="leading-loose">
+            <span class="capitalize font-secondary font-bold"> Archivo Narrow bold </span>
+            <code>.font-bold</code>
+          </span>
+        </li>
+        <li></li>
+      </ol>
+    </div>
+  </div>
+</template>
