+++ packages/vue/src/docs/foundation/color.docs.mdx
