+++ packages/vue/src/templates/PageContent/PageContent.stories.js
@@ -1,11 +1,11 @@
-import { BlockHeadingData } from './../components/BlockHeading/BlockHeading.stories'
-import { BlockKeyPointsData } from './../components/BlockKeyPoints/BlockKeyPoints.stories'
-import { BlockImageComparisonData } from './../components/BlockImageComparison/BlockImageComparison.stories'
-import { BlockIframeEmbedData } from './../components/BlockIframeEmbed/BlockIframeEmbed.stories'
-import { HeroMediaData } from './../components/HeroMedia/HeroMedia.stories'
-import { BlockRelatedLinksData } from './../components/BlockRelatedLinks/BlockRelatedLinks.stories.js'
-import { BlockLinkCardCarouselData } from './../components/BlockLinkCarousel/BlockLinkCarousel.stories.js'
-import { NavSecondaryData } from './../components/NavSecondary/NavSecondary.stories.js'
+import { BlockHeadingData } from './../../components/BlockHeading/BlockHeading.stories'
+import { BlockKeyPointsData } from './../../components/BlockKeyPoints/BlockKeyPoints.stories'
+import { BlockImageComparisonData } from './../../components/BlockImageComparison/BlockImageComparison.stories'
+import { BlockIframeEmbedData } from './../../components/BlockIframeEmbed/BlockIframeEmbed.stories'
+import { HeroMediaData } from './../../components/HeroMedia/HeroMedia.stories'
+import { BlockRelatedLinksData } from './../../components/BlockRelatedLinks/BlockRelatedLinks.stories.js'
+import { BlockLinkCardCarouselData } from './../../components/BlockLinkCarousel/BlockLinkCarousel.stories.js'
+import { NavSecondaryData } from './../../components/NavSecondary/NavSecondary.stories.js'
 
 import PageContent from './PageContent.vue'
